--- lib/base_class.rb
@@ -1,6 +1,0 @@
-class BaseClass
-  CAMPAIGN_TYPE_CODES = {
-    sp: 'sp',
-    sb: 'hsa'
-  }.freeze
-end
