+++ lib/blurb/base_class.rb
@@ -1,0 +1,8 @@
+class Blurb
+  class BaseClass
+    CAMPAIGN_TYPE_CODES = {
+      sp: 'sp',
+      sb: 'hsa'
+    }.freeze
+  end
+end
